requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
sqlalchemy
aiosqlite
pytest
pytest-asyncio
greenlet